1. Ability to provide orientdb configuration  and benchmark properties through property file was added.   2. Overflow error was fixed.

+++ src/main/java/com/yahoo/ycsb/DBWrapper.java
@@ -164,9 +164,9 @@ public class DBWrapper extends DB
       }
     }
     _measurements.measure(measurementName,
-        (int)((endTimeNanos-startTimeNanos)/1000));
+        ((endTimeNanos-startTimeNanos)/1000));
     _measurements.measureIntended(measurementName,
-        (int)((endTimeNanos-intendedStartTimeNanos)/1000));
+        ((endTimeNanos-intendedStartTimeNanos)/1000));
   }
 
   /**
