+++ src/main/java/com/yahoo/ycsb/measurements/OneMeasurement.java
@@ -45,7 +45,7 @@ public abstract class OneMeasurement {
     this._returncodes = new ConcurrentHashMap<Status, AtomicInteger>();
   }
 
-  public abstract void measure(int latency);
+  public abstract void measure(long latency);
 
   public abstract String getSummary();
